+++ core/build.gradle.kts
@@ -2,20 +2,9 @@ plugins {
     id("benchmark.kotlin-conventions")
 }
 
-group = "com.github.jschlicht"
-version = "1.0-SNAPSHOT"
-
-repositories {
-    mavenCentral()
-}
-
 dependencies {
-    testImplementation("org.jetbrains.kotlin:kotlin-test")
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-kotlin {
-    jvmToolchain(21)
+    implementation(libs.jooq.jooq)
+    implementation(libs.testcontainers.mariadb)
+    implementation(libs.testcontainers.mysql)
+    implementation(libs.testcontainers.postgres)
 }
