+++ core/src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/core/db/Postgres.kt
@@ -1,4 +1,4 @@
-package com.github.jschlicht.multitenantdbbenchmark.db
+package com.github.jschlicht.multitenantdbbenchmark.core.db
 
 import org.testcontainers.utility.DockerImageName
