+++ core/src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/core/strategy/DistributedTable.kt
@@ -1,4 +1,4 @@
-package com.github.jschlicht.multitenantdbbenchmark.strategy
+package com.github.jschlicht.multitenantdbbenchmark.core.strategy
 
 data object DistributedTable : Strategy("distributed_table") {
     override val simpleKeys = false
