+++ core/src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/core/strategy/PartitionHash.kt
@@ -1,4 +1,4 @@
-package com.github.jschlicht.multitenantdbbenchmark.strategy
+package com.github.jschlicht.multitenantdbbenchmark.core.strategy
 
 data object PartitionHash : Strategy("partition_hash") {
     override val simpleKeys = false
