+++ core/src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/core/strategy/PartitionList.kt
@@ -1,4 +1,4 @@
-package com.github.jschlicht.multitenantdbbenchmark.strategy
+package com.github.jschlicht.multitenantdbbenchmark.core.strategy
 
 data object PartitionList : Strategy("partition_list") {
     override val simpleKeys = false
