+++ core/src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/core/strategy/TenantIdComposite.kt
@@ -1,4 +1,4 @@
-package com.github.jschlicht.multitenantdbbenchmark.strategy
+package com.github.jschlicht.multitenantdbbenchmark.core.strategy
 
 data object TenantIdComposite : Strategy("tenant_id_composite") {
     override val simpleKeys = false
