+++ core/src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/core/strategy/TenantIdSimple.kt
@@ -1,4 +1,4 @@
-package com.github.jschlicht.multitenantdbbenchmark.strategy
+package com.github.jschlicht.multitenantdbbenchmark.core.strategy
 
 data object TenantIdSimple : Strategy("tenant_id_simple") {
     override val simpleKeys = true
