+++ data/build.gradle.kts
@@ -1,21 +1,3 @@
 plugins {
     id("benchmark.kotlin-conventions")
 }
-
-group = "com.github.jschlicht"
-version = "1.0-SNAPSHOT"
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation("org.jetbrains.kotlin:kotlin-test")
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-kotlin {
-    jvmToolchain(21)
-}
