+++ src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/Main.kt
@@ -4,8 +4,8 @@ import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.options.*
 import com.github.ajalt.clikt.parameters.types.choice
 import com.github.ajalt.clikt.parameters.types.path
-import com.github.jschlicht.multitenantdbbenchmark.db.*
-import com.github.jschlicht.multitenantdbbenchmark.strategy.*
+import com.github.jschlicht.multitenantdbbenchmark.core.db.*
+import com.github.jschlicht.multitenantdbbenchmark.core.strategy.*
 import com.github.jschlicht.multitenantdbbenchmark.util.MdcKey
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.github.oshai.kotlinlogging.withLoggingContext
