+++ src/main/kotlin/com/github/jschlicht/multitenantdbbenchmark/SuiteContext.kt
@@ -1,6 +1,6 @@
 package com.github.jschlicht.multitenantdbbenchmark
 
-import com.github.jschlicht.multitenantdbbenchmark.db.Database
-import com.github.jschlicht.multitenantdbbenchmark.strategy.Strategy
+import com.github.jschlicht.multitenantdbbenchmark.core.db.Database
+import com.github.jschlicht.multitenantdbbenchmark.core.strategy.Strategy
 
 data class SuiteContext(val databases: List<Database>, val strategies: List<Strategy>)
